Prepare for RouteResultObserver updates
As of zendframework/zend-expressive#200, a new interface,
`Zend\Expressive\Router\RouteResultObserverInterface` exists to allow notifying
instances of the calculated `RouteResult`; this is particularly of interest to
URI generators.

With zendframework/zend-expressive-zendviewrenderer#9, the zend-view integration
provides an observer via the `UrlHelper` it ships, and allows you to attach it
to the application via a delegator factory.

This patch defines the delegator factory in the template configuration for
zend-view, and also details how you would accomplish similar functionality with
Pimple.