Updated serve command and Apache asset rules
The changes in this patch were introduced after testing several debug bars.

The `serve` command was updated to include the script argument
`public/index.php`. This was done to ensure that the script intercepts for every
request; if omitted, requests that look like filenames are returned as 404s by
the built-in web server. (This caused issues with debug bars that serve assets
via streams.)

The Apache `.htaccess` rules were updated to remove a section that was
specifically matching web assets by extension (including CSS, JS, fonts, and
most image formats). This rule was removed for the following reasons:

- The preceding rule would prevent it from triggering in most cases already, as
  it checks to see if the request is for a file on the filesystem, and, if so,
  returns it immediately.
- If the preceding rules failed, we likely want to serve the request through the
  web application (e.g., via Assetic, vfs, etc.). The rules as written would
  prevent that from happening.