Added clear-config-cache utility
Per conversation on #54, we want to make production settings the
default, which means enabling the configuration cache. However, we also
need to make it easy for developers to clear the configuration cache.

This patch adds a script, `bin/clear-config-cache.php`, which can be
used to remove the configured configuration cache file, if defined. It
is aliased in `composer.json` as `clear-config-cache`, allowing usage as
`composer clear-config-cache`.