Provide service aliasing support to Aura.Di and Pimple
With the addition of aliases for the `DefaultDelegate` service, we need
to support aliasing for all supported containers. This patch adds that
support for both Pimple and Aura.Di. In the case of pimple, this means
simply setting the alias service to return the value of fetching the
target service. In the case of Aura.Di, I used an approach covered in:

- http://thoughts.silentworks.co.uk/aura-di-service-aliasing