Update to zend-expressive-helpers 4.0
zend-expressive-helpers 4.0 updates to PSR-11 and PSR-15 interfaces,
and, as such, is backwards incompatible with previous releases for
purposes of extension. For purposes of consumption, which is what the
skeleton does, it is fully backwards compatible.

Due to the BC breaks, new minor releases of each renderer were made that
add support for the helpers 4.0 series.

As such, this patch updates to use zend-expressive-helpers 4.0, and to
install the renderer versions compatible with that release.

Additionally

- it updates the zend-expressive-tooling minimum supported version to
  pick up the latest bugfixes.
- it updates the zend-expressive minimum supported version to pick up
  the latest bugfixes.