Have pipeline and route files return closures
This patch implements an idea [proposed by @Xerkus](https://github.com/zendframework/zend-expressive/pull/543#issuecomment-361419425):
having the `pipeline.php` and `routes.php` files return _closures_, and
updating `public/index.php` to capitalize on that in order to pass in
arguments.

In this particular case, we pass the application, middleware factory,
and PSR-11 container instances as arguments. This provides developers
with all the tools they should need in order to create the pipeline
and/or routes.