Reverts usage of ROUTE_MIDDLEWARE constant
This patch follows on zendframework/zend-expressive#559, and modifies
the pipeline to directly refer to the `PathBasedRoutingMiddleware`
instead of using the `ROUTE_MIDDLEWARE` constant. It's better and faster
to use the canonical name, and the old name (to which the constant
resolves) is an alias for purposes of aiding those upgrading from v2;
since aliases resolve to the same instance as the canonical service,
this remains backwards compatible.