Use zend-expressive-router 3.0.0rc4
This patch updates the skeleton to use zend-expressive-router 3.0.0rc4,
as well as the compatible versions of each of the existing adapters.

One change necessary is for the pipeline to change the order of the
various zend-expressive-router middleare;
`MethodNotAllowedMiddleware` MUST follow the `Implicit*Middleware` to
ensure it does not respond for OPTIONS and HEAD requests.