Fixes issue with support section of composer.json
`composer validate` indicates that any IRC links need to use the
`irc://` protocol, which ours did not.

As it turns out, the json schema allows _arbitrary_ link types; it only
defines constraints for _known_ types, such as IRC.

As such, I changed this back to "slack", as it better indicates exactly
the type of support channel it represents.